# backend/generated/openapi/model_apartment_avialability_response.go
/*
TAPO Apartments API

API for TAPO Apartments website

API version: 1.0.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package openapi

import (
	"encoding/json"
	"time"
	"bytes"
	"fmt"
)

// checks if the ApartmentAvialabilityResponse type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &ApartmentAvialabilityResponse{}

// ApartmentAvialabilityResponse struct for ApartmentAvialabilityResponse
type ApartmentAvialabilityResponse struct {
	// Apartment slug
	Slug string `json:"slug"`
	// List of external unavailable date ranges
	ExternalUnavailableDates []UnavailableDateRange `json:"externalUnavailableDates"`
	// List of internal unavailable date ranges
	InternalUnavailableDates []UnavailableDateRange `json:"internalUnavailableDates"`
	// Date of the last sync with external system
	ExternalLastSyncedAt time.Time `json:"externalLastSyncedAt"`
}

type _ApartmentAvialabilityResponse ApartmentAvialabilityResponse

// NewApartmentAvialabilityResponse instantiates a new ApartmentAvialabilityResponse object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewApartmentAvialabilityResponse(slug string, externalUnavailableDates []UnavailableDateRange, internalUnavailableDates []UnavailableDateRange, externalLastSyncedAt time.Time) *ApartmentAvialabilityResponse {
	this := ApartmentAvialabilityResponse{}
	this.Slug = slug
	this.ExternalUnavailableDates = externalUnavailableDates
	this.InternalUnavailableDates = internalUnavailableDates
	this.ExternalLastSyncedAt = externalLastSyncedAt
	return &this
}

// NewApartmentAvialabilityResponseWithDefaults instantiates a new ApartmentAvialabilityResponse object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewApartmentAvialabilityResponseWithDefaults() *ApartmentAvialabilityResponse {
	this := ApartmentAvialabilityResponse{}
	return &this
}

// GetSlug returns the Slug field value
func (o *ApartmentAvialabilityResponse) GetSlug() string {
	if o == nil {
		var ret string
		return ret
	}

	return o.Slug
}

// GetSlugOk returns a tuple with the Slug field value
// and a boolean to check if the value has been set.
func (o *ApartmentAvialabilityResponse) GetSlugOk() (*string, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Slug, true
}

// SetSlug sets field value
func (o *ApartmentAvialabilityResponse) SetSlug(v string) {
	o.Slug = v
}

// GetExternalUnavailableDates returns the ExternalUnavailableDates field value
func (o *ApartmentAvialabilityResponse) GetExternalUnavailableDates() []UnavailableDateRange {
	if o == nil {
		var ret []UnavailableDateRange
		return ret
	}

	return o.ExternalUnavailableDates
}

// GetExternalUnavailableDatesOk returns a tuple with the ExternalUnavailableDates field value
// and a boolean to check if the value has been set.
func (o *ApartmentAvialabilityResponse) GetExternalUnavailableDatesOk() ([]UnavailableDateRange, bool) {
	if o == nil {
		return nil, false
	}
	return o.ExternalUnavailableDates, true
}

// SetExternalUnavailableDates sets field value
func (o *ApartmentAvialabilityResponse) SetExternalUnavailableDates(v []UnavailableDateRange) {
	o.ExternalUnavailableDates = v
}

// GetInternalUnavailableDates returns the InternalUnavailableDates field value
func (o *ApartmentAvialabilityResponse) GetInternalUnavailableDates() []UnavailableDateRange {
	if o == nil {
		var ret []UnavailableDateRange
		return ret
	}

	return o.InternalUnavailableDates
}

// GetInternalUnavailableDatesOk returns a tuple with the InternalUnavailableDates field value
// and a boolean to check if the value has been set.
func (o *ApartmentAvialabilityResponse) GetInternalUnavailableDatesOk() ([]UnavailableDateRange, bool) {
	if o == nil {
		return nil, false
	}
	return o.InternalUnavailableDates, true
}

// SetInternalUnavailableDates sets field value
func (o *ApartmentAvialabilityResponse) SetInternalUnavailableDates(v []UnavailableDateRange) {
	o.InternalUnavailableDates = v
}

// GetExternalLastSyncedAt returns the ExternalLastSyncedAt field value
func (o *ApartmentAvialabilityResponse) GetExternalLastSyncedAt() time.Time {
	if o == nil {
		var ret time.Time
		return ret
	}

	return o.ExternalLastSyncedAt
}

// GetExternalLastSyncedAtOk returns a tuple with the ExternalLastSyncedAt field value
// and a boolean to check if the value has been set.
func (o *ApartmentAvialabilityResponse) GetExternalLastSyncedAtOk() (*time.Time, bool) {
	if o == nil {
		return nil, false
	}
	return &o.ExternalLastSyncedAt, true
}

// SetExternalLastSyncedAt sets field value
func (o *ApartmentAvialabilityResponse) SetExternalLastSyncedAt(v time.Time) {
	o.ExternalLastSyncedAt = v
}

func (o ApartmentAvialabilityResponse) MarshalJSON() ([]byte, error) {
	toSerialize,err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o ApartmentAvialabilityResponse) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	toSerialize["slug"] = o.Slug
	toSerialize["externalUnavailableDates"] = o.ExternalUnavailableDates
	toSerialize["internalUnavailableDates"] = o.InternalUnavailableDates
	toSerialize["externalLastSyncedAt"] = o.ExternalLastSyncedAt
	return toSerialize, nil
}

func (o *ApartmentAvialabilityResponse) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"slug",
		"externalUnavailableDates",
		"internalUnavailableDates",
		"externalLastSyncedAt",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err;
	}

	for _, requiredProperty := range(requiredProperties) {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varApartmentAvialabilityResponse := _ApartmentAvialabilityResponse{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varApartmentAvialabilityResponse)

	if err != nil {
		return err
	}

	*o = ApartmentAvialabilityResponse(varApartmentAvialabilityResponse)

	return err
}

type NullableApartmentAvialabilityResponse struct {
	value *ApartmentAvialabilityResponse
	isSet bool
}

func (v NullableApartmentAvialabilityResponse) Get() *ApartmentAvialabilityResponse {
	return v.value
}

func (v *NullableApartmentAvialabilityResponse) Set(val *ApartmentAvialabilityResponse) {
	v.value = val
	v.isSet = true
}

func (v NullableApartmentAvialabilityResponse) IsSet() bool {
	return v.isSet
}

func (v *NullableApartmentAvialabilityResponse) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableApartmentAvialabilityResponse(val *ApartmentAvialabilityResponse) *NullableApartmentAvialabilityResponse {
	return &NullableApartmentAvialabilityResponse{value: val, isSet: true}
}

func (v NullableApartmentAvialabilityResponse) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableApartmentAvialabilityResponse) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}